feat: make LoadHTTPConfigFile set directory and move from tests file
